feat: execute scarb build
although the issue description said to work in the src/proof_generator dir. i couldn't locate such, so i added a file in the proof_client folder.

to test this, you can run cargo test --test scarb_build.

the test basically creates a temporary project that is dropped once the function is not in scope any more.