feat: switch L1 listener to single DepositEvent
i also ensure that the commitment_hash is unique by ensuring that is indexed in the deposits table (i included the required migration) so now, no two deposits can have the commitment hash.

this also instroduces a retry/backoff mechanism in the fetch_events_logs_at_address function. for now, the max retries is 5. we can update this, if we need to increase though.